stable_diffusion/linux_requirements.txt
pynecone==0.1.33
torch
diffusers
pillow
transformers
accelerate
xformers